// src/types/history.ts
export interface History {
  testTaker: string
  date: string
  completion: string
  payment: string
  viewProfile: string
  email: string
  docId: string
  userId: string
  certificateNo: string
  mailSent: boolean
}

export interface UserHistory {
  date: string
  score: string
  language: string
  certificateNo: string
  payment: string
  docId: string
}